add hateoas links

+++ app/controllers/user.js
@@ -1,9 +1,11 @@
 const bcrypt = require("bcrypt");
 const jwt = require("jsonwebtoken");
 const User = require("../models/user");
+const Sauce = require("../models/sauce");
 //import cryptojs for encrypt email
 const CryptoJS = require("crypto-js");
 const { json } = require("express");
+const fs = require("fs");
 
 require("dotenv").config();
 
@@ -47,7 +49,12 @@ exports.signup = (req, res, next) => {
         .save()
         .then((newUser) => {
           newUser.email = decrypt(newUser.email);
-          res.status(201).json({ message: "User created !", user: newUser });
+          res
+            .status(201)
+            .json(
+              { message: "User created !", user: newUser },
+              hateoasLinks(req)
+            );
         })
         .catch((error) => res.status(400).json({ error }));
     })
@@ -73,19 +80,22 @@ exports.login = (req, res, next) => {
               .status(401)
               .json({ error: "Your password is incorrect !" });
           }
-          res.status(200).json({
-            userId: user._id,
-            //chiffrer un nouveau token
-            token: jwt.sign(
-              //userId entant playload
-              { userId: user._id },
-              //random token dispo pendant 24h
-              process.env.JWT_TOKEN,
-              { expiresIn: "24h" }
-            ),
-            //return user as correct user
-            user: user,
-          });
+          res.status(200).json(
+            {
+              userId: user._id,
+              //chiffrer un nouveau token
+              token: jwt.sign(
+                //userId entant playload
+                { userId: user._id },
+                //random token dispo pendant 24h
+                process.env.JWT_TOKEN,
+                { expiresIn: "24h" }
+              ),
+              //return user as correct user
+              user: user,
+            },
+            hateoasLinks(req, user._id)
+          );
         })
         .catch((error) => res.status(500).json({ error }));
     })
@@ -102,7 +112,7 @@ exports.readUser = (req, res, next) => {
       } else {
         // decrypt the email to be returned
         user.email = decrypt(user.email);
-        res.status(200).json({ user });
+        res.status(200).json({ user }, hateoasLinks(req, user._id));
       }
     })
     .catch((error) => res.status(500).json(error));
@@ -144,7 +154,11 @@ exports.updateUser = (req, res, next) => {
             updatedUser.email = decrypt(updatedUser.email);
             res
               .status(200)
-              .json({ message: "User has been updated", updatedUser });
+              .json(
+                { message: "User has been updated" },
+                updatedUser,
+                hateoasLinks(req, this.updateUser._id)
+              );
           })
           .catch((error) => res.status(400).json(error));
       }
@@ -154,9 +168,56 @@ exports.updateUser = (req, res, next) => {
 
 //delete account
 exports.deleteUser = (req, res, next) => {
-  User.deleteOne({ _id: req.auth.userId })
-    .then(() => {
-      res.status(204).json({ message: "User deleted" });
+  User.findById(req.params.userId)
+    // check the email of user
+    .then((user) => {
+      if (!user) {
+        res.status(401).json({ message: "user not found" });
+      } else {
+        fs.unlink(`images`, () => {});
+        User.deleteOne({ _id: req.auth.userId })
+          .then(() => {
+            res.status(204).json({ message: "User deleted" }, hateoasLinks());
+          })
+          .catch((error) => ({ error }));
+      }
     })
-    .catch((error) => ({ error }));
+    .catch((error) => {
+      console.log(error);
+    });
+};
+
+const hateoasLinks = () => {
+  return [
+    ({
+      href: `${req.protocol}://${req.get("host") + "/api/auth/signup"}`,
+      rel: "signup",
+      type: "POST",
+    },
+    {
+      href: `${req.protocol}://${req.get("host") + "/api/auth/login"}`,
+      rel: "login",
+      type: "POST",
+    },
+    {
+      href: `${req.protocol}://${req.get("host") + "/api/auth/"}`,
+      rel: "read",
+      type: "GET",
+    },
+    {
+      href: `${req.protocol}://${req.get("host") + "/api/auth/export"}`,
+      rel: "export",
+      type: "GET",
+    },
+    {
+      href: `${req.protocol}://${req.get("host") + "/api/auth/"}`,
+      rel: "update",
+      type: "PUT",
+    },
+    {
+      href: `${req.protocol}://${req.get("host") + "/api/auth/"}`,
+      rel: "delete",
+      type: "DELETE",
+    }),
+  ];
 };
